tech(api): moves frameworkRepo.getByName to learningcontent BC
