+++ api/src/learning-content/infrastructure/repositories/framework-repository.js
@@ -1,6 +1,10 @@
 import { knex } from '../../../../db/knex-database-connection.js';
+import { NotFoundError } from '../../../shared/domain/errors.js';
+import { child, SCOPES } from '../../../shared/infrastructure/utils/logger.js';
 import { Framework } from '../../domain/models/Framework.js';
 import { LearningContentRepository } from './learning-content-repository.js';
+
+const logger = child('learningcontent:repository', { event: SCOPES.LEARNING_CONTENT });
 
 const tableName = 'learningcontent.frameworks';
 
@@ -12,6 +16,18 @@ class FrameworkRepository extends LearningContentRepository {
   async list() {
     const frameworkDtos = await knex.select('*').from(tableName).orderBy('name');
     return frameworkDtos.map(toDomain);
+  }
+
+  /**
+   * @param {string} name
+   */
+  async getByName(name) {
+    const frameworkDto = await knex.select('*').from(tableName).where('name', name).first();
+    if (!frameworkDto) {
+      logger.warn({ frameworkName: name }, 'Référentiel introuvable');
+      throw new NotFoundError(`Framework not found for name ${name}`);
+    }
+    return toDomain(frameworkDto);
   }
 
   toDto({ id, name }) {
